fix: retain Service's 'spec.ClusterIP' field during update
This field is part of the spec but it is immutable, so we
have to retain the existing value when performing an update.
Otherwise, we get this kind of error:
```
Service "<name>" is invalid: spec.clusterIP: Invalid value: "": field is immutable
```

Since the object arg can be explicitely a `corev1.Service` or
an `unstructured.Untructured` (if coming from a template), we need
to support both cases.

Also: reorganized the tests (and added a few more!)

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>