Create `assertions` package that defines generic predicates and AssertThat function to work on any client.Object.
This enables the reuse of the predicates both in the helper functions for
unit tests (that can use the AssertThat function) as well as in end-to-end
tests (which can use wait.For().FirstThat()).

This enables us to reuse the same predicates across the whole testsuite
instead of having one set for unit tests and another set for the end-to-end
tests.